+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @hypersonic-js/cli
 
+## 0.2.5
+
+### Patch Changes
+
+- Bump deps in cli templates
+- Updated dependencies
+  - @hypersonic-js/admin@0.2.5
+
 ## 0.2.4
 
 ### Patch Changes